+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Unreleased
 
 - Update to Kotlin `2.1.21`.
 - Build against KSP `2.1.21-2.0.1`.
+- Fix not being able to provide a custom ViewModel to `continuityRetainedStateRegistry()`
 
 0.28.0
 ------